appstuff/requirements.txt
Flask==2.3.0
flask-socketio==5.3.0
gevent==22.10.2
gunicorn==20.1.0
discord.py==2.0.1
requests==2.28.1